refactor : #127 Review 서비스 구조 리팩터링
