+++ com.taken_seat.review-service/src/test/java/com/taken_seat/review_service/unit/ReviewServicesTest.java
@@ -27,15 +27,15 @@ import com.taken_seat.review_service.application.dto.request.ReviewRegisterReqDt
 import com.taken_seat.review_service.application.dto.request.ReviewUpdateReqDto;
 import com.taken_seat.review_service.application.dto.response.PageReviewResponseDto;
 import com.taken_seat.review_service.application.dto.response.ReviewDetailResDto;
-import com.taken_seat.review_service.application.service.ReviewService;
 import com.taken_seat.review_service.domain.model.Review;
 import com.taken_seat.review_service.domain.repository.ReviewQuerydslRepository;
 import com.taken_seat.review_service.domain.repository.ReviewRepository;
 import com.taken_seat.review_service.infrastructure.client.dto.BookingStatusDto;
 import com.taken_seat.review_service.infrastructure.client.dto.PerformanceEndTimeDto;
+import com.taken_seat.review_service.infrastructure.service.ReviewServiceImpl;
 
 @ExtendWith(MockitoExtension.class)
-public class ReviewServiceTest {
+public class ReviewServicesTest {
 
 	@Mock
 	private ReviewRepository reviewRepository;
@@ -47,7 +47,7 @@ public class ReviewServiceTest {
 	private ReviewClient reviewClient;
 
 	@InjectMocks
-	private ReviewService reviewService;
+	private ReviewServiceImpl reviewServices;
 
 	private UUID testPerformanceId;
 
@@ -104,7 +104,7 @@ public class ReviewServiceTest {
 		when(reviewRepository.save(any(Review.class))).thenAnswer(invocation -> invocation.getArgument(0));
 
 		// When
-		ReviewDetailResDto result = reviewService.registerReview(requestDto, authenticatedUser);
+		ReviewDetailResDto result = reviewServices.registerReview(requestDto, authenticatedUser);
 
 		// Then
 		assertNotNull(result);
@@ -125,7 +125,7 @@ public class ReviewServiceTest {
 
 		// When & Then
 		ReviewException ex = assertThrows(ReviewException.class, () -> {
-			reviewService.registerReview(requestDto, authenticatedUser);
+			reviewServices.registerReview(requestDto, authenticatedUser);
 		});
 
 		assertEquals(ResponseCode.REVIEW_ALREADY_WRITTEN.getMessage(), ex.getMessage());
@@ -144,7 +144,7 @@ public class ReviewServiceTest {
 
 		// When & Then
 		ReviewException ex = assertThrows(ReviewException.class, () -> {
-			reviewService.registerReview(requestDto, authenticatedUser);
+			reviewServices.registerReview(requestDto, authenticatedUser);
 		});
 
 		assertEquals(ResponseCode.BOOKING_NOT_COMPLETED.getMessage(), ex.getMessage());
@@ -165,7 +165,7 @@ public class ReviewServiceTest {
 
 		// When & Then
 		ReviewException ex = assertThrows(ReviewException.class, () -> {
-			reviewService.registerReview(requestDto, authenticatedUser);
+			reviewServices.registerReview(requestDto, authenticatedUser);
 		});
 
 		assertEquals(ResponseCode.EARLY_REVIEW.getMessage(), ex.getMessage());
@@ -178,7 +178,7 @@ public class ReviewServiceTest {
 		when(reviewRepository.findByIdAndDeletedAtIsNull(testReviewId)).thenReturn(Optional.of(testReview));
 
 		// When
-		ReviewDetailResDto result = reviewService.getReviewDetail(testReviewId);
+		ReviewDetailResDto result = reviewServices.getReviewDetail(testReviewId);
 
 		// Then
 		assertNotNull(result);
@@ -196,7 +196,7 @@ public class ReviewServiceTest {
 
 		// When & Then
 		ReviewException ex = assertThrows(ReviewException.class, () -> {
-			reviewService.getReviewDetail(TestRegisterReviewId);
+			reviewServices.getReviewDetail(TestRegisterReviewId);
 		});
 
 		assertEquals(ResponseCode.REVIEW_NOT_FOUND.getMessage(), ex.getMessage());
@@ -220,7 +220,7 @@ public class ReviewServiceTest {
 		when(reviewQuerydslRepository.search(query, category, page, size, sort, order)).thenReturn(mockPage);
 
 		// When
-		PageReviewResponseDto result = reviewService.searchReview(query, category, page, size, sort, order);
+		PageReviewResponseDto result = reviewServices.searchReview(query, category, page, size, sort, order);
 
 		// Then
 		assertNotNull(result);
@@ -241,7 +241,7 @@ public class ReviewServiceTest {
 			.thenReturn(Page.empty());
 
 		// When
-		PageReviewResponseDto result = reviewService.searchReview("emptyTitle", "title", 0, 10, "createdAt", "desc");
+		PageReviewResponseDto result = reviewServices.searchReview("emptyTitle", "title", 0, 10, "createdAt", "desc");
 
 		// Then
 		assertNotNull(result);
@@ -259,7 +259,7 @@ public class ReviewServiceTest {
 		when(reviewRepository.findByIdAndDeletedAtIsNull(testReviewId)).thenReturn(Optional.of(testReview));
 
 		// When
-		ReviewDetailResDto result = reviewService.updateReview(testReviewId, reqDto, authenticatedUser);
+		ReviewDetailResDto result = reviewServices.updateReview(testReviewId, reqDto, authenticatedUser);
 
 		// Then
 		assertNotNull(result);
@@ -279,7 +279,7 @@ public class ReviewServiceTest {
 
 		// When & Then
 		ReviewException exception = assertThrows(ReviewException.class, () -> {
-			reviewService.updateReview(notExistId, reqDto, authenticatedUser);
+			reviewServices.updateReview(notExistId, reqDto, authenticatedUser);
 		});
 
 		assertEquals("해당 리뷰가 존재하지않습니다.", exception.getMessage());
@@ -297,7 +297,7 @@ public class ReviewServiceTest {
 
 		// When & Then
 		ReviewException exception = assertThrows(ReviewException.class, () -> {
-			reviewService.updateReview(testReviewId, reqDto, anotherUser);
+			reviewServices.updateReview(testReviewId, reqDto, anotherUser);
 		});
 
 		assertEquals("해당 리뷰에 접근할 권한이 없습니다.", exception.getMessage()); // FORBIDDEN_REVIEW_ACCESS 메시지
@@ -310,7 +310,7 @@ public class ReviewServiceTest {
 		when(reviewRepository.findByIdAndDeletedAtIsNull(testReviewId)).thenReturn(Optional.of(testReview));
 
 		// When
-		assertDoesNotThrow(() -> reviewService.deleteReview(testReviewId, authenticatedUser));
+		assertDoesNotThrow(() -> reviewServices.deleteReview(testReviewId, authenticatedUser));
 
 		// Then
 		assertNotNull(testReview.getDeletedAt());
@@ -326,7 +326,7 @@ public class ReviewServiceTest {
 		AuthenticatedUser anotherUser = new AuthenticatedUser(notAuthorId, "other@gmail.com", "other@gmail.com");
 		// When & Then
 		ReviewException exception = assertThrows(ReviewException.class, () ->
-			reviewService.deleteReview(UUID.randomUUID(), anotherUser)
+			reviewServices.deleteReview(UUID.randomUUID(), anotherUser)
 		);
 
 		assertEquals("해당 리뷰가 존재하지않습니다.", exception.getMessage());
@@ -343,7 +343,7 @@ public class ReviewServiceTest {
 
 		// When & Then
 		ReviewException exception = assertThrows(ReviewException.class, () -> {
-			reviewService.deleteReview(testReviewId, anotherUser);
+			reviewServices.deleteReview(testReviewId, anotherUser);
 		});
 
 		assertEquals("해당 리뷰에 접근할 권한이 없습니다.", exception.getMessage()); // FORBIDDEN_REVIEW_ACCESS 메시지
